Raise an exception and exit with a meaningful message when binary path substitution fails.

+++ lib/spack/spack/relocate.py
@@ -28,6 +28,19 @@ class InstallRootStringException(spack.error.SpackError):
             (file_path, root_path))
 
 
+class BinaryStringReplacementException(spack.error.SpackError):
+    """
+    Raised when the size of the file changes after binary path substitution.
+    """
+
+    def __init__(self, file_path, old_len, new_len):
+        super(BinaryStringReplacementException, self).__init__(
+            "Doing a binary string replacement in %s failed.\n"
+            "The size of the file changed from %s to %s\n"
+            "when it should have remanined the same." %
+            (file_path, old_len, new_len))
+
+
 def get_patchelf():
     """
     Builds and installs spack patchelf package on linux platforms
@@ -308,8 +321,11 @@ def replace_prefix_bin(path_name, old_dir, new_dir):
         f.seek(0)
         original_data_len = len(data)
         pat = re.compile(re.escape(old_dir) + b'([^\0]*?)\0')
-        data = pat.sub(replace, data)
-        assert(len(data) == original_data_len)
+        ndata = pat.sub(replace, data)
+        new_data_len = len(ndata)
+        if not new_data_len == original_data_len:
+            raise BinaryStringReplacementException(
+                path_name, original_data_len, new_data_len)
         f.write(data)
         f.truncate()
